refactor: extract validation traits to the same module
