+++ crates/amaru-ouroboros-traits/src/validators/blocks/can_validate_blocks.rs
@@ -1,4 +1,4 @@
-// Copyright 2025 PRAGMA
+// Copyright 2026 PRAGMA
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -14,12 +14,10 @@
 
 use std::fmt::{Debug, Display, Formatter};
 
-use amaru_kernel::{Block, BlockHeader, Point, Tip};
+use amaru_kernel::{Block, Point, Tip};
 use amaru_metrics::ledger::LedgerMetrics;
 use serde::{Deserialize, Serialize};
 use thiserror::Error;
-
-pub mod mock;
 
 #[async_trait::async_trait]
 pub trait CanValidateBlocks: Send + Sync {
@@ -95,71 +93,6 @@ impl<'de> Deserialize<'de> for BlockValidationError {
 }
 
 impl PartialEq for BlockValidationError {
-    fn eq(&self, other: &Self) -> bool {
-        self.0.to_string() == other.0.to_string()
-    }
-}
-
-pub trait CanValidateHeaders: Send + Sync {
-    fn validate_header(&self, header: &BlockHeader) -> Result<(), HeaderValidationError>;
-}
-
-#[derive(Debug)]
-pub struct HeaderValidationError(anyhow::Error);
-
-impl HeaderValidationError {
-    pub fn new(err: anyhow::Error) -> Self {
-        HeaderValidationError(err)
-    }
-
-    pub fn to_anyhow(self) -> anyhow::Error {
-        self.0
-    }
-
-    pub fn downcast<T: std::error::Error + Debug + Send + Sync + 'static>(self) -> Result<T, anyhow::Error> {
-        self.0.downcast::<T>()
-    }
-
-    pub fn downcast_ref<T: std::error::Error + Debug + Send + Sync + 'static>(&self) -> Option<&T> {
-        self.0.downcast_ref::<T>()
-    }
-}
-
-impl From<anyhow::Error> for HeaderValidationError {
-    fn from(err: anyhow::Error) -> Self {
-        HeaderValidationError::new(err)
-    }
-}
-
-impl Display for HeaderValidationError {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        write!(f, "HeaderValidationError: {}", self.0)
-    }
-}
-
-impl Serialize for HeaderValidationError {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        serializer.serialize_str(&self.0.to_string())
-    }
-}
-
-/// This deserialization implementation is a best-effort attempt to
-/// recover the error message. The original error type is lost during
-/// serialization, so we can only reconstruct the error message as a string.
-impl<'de> Deserialize<'de> for HeaderValidationError {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: serde::Deserializer<'de>,
-    {
-        let s = String::deserialize(deserializer)?;
-        Ok(HeaderValidationError::new(anyhow::anyhow!(s)))
-    }
-}
-
-impl PartialEq for HeaderValidationError {
     fn eq(&self, other: &Self) -> bool {
         self.0.to_string() == other.0.to_string()
     }
