+++ crates/amaru-ouroboros-traits/src/validators/headers/mod.rs
@@ -1,4 +1,4 @@
-// Copyright 2025 PRAGMA
+// Copyright 2026 PRAGMA
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,20 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub mod has_stake_distribution;
-pub use has_stake_distribution::{HasStakeDistribution, PoolSummary};
+pub mod can_validate_headers;
+pub mod can_validate_headers_mock;
 
-pub mod validators;
-pub use validators::*;
-
-pub mod connections;
-pub use connections::*;
-
-pub mod mempool;
-pub use mempool::*;
-
-pub mod stores;
-pub use stores::*;
-
-pub mod praos;
-pub use praos::*;
+pub use can_validate_headers::*;
+pub use can_validate_headers_mock::*;
