+++ crates/amaru-ouroboros-traits/src/validators/mod.rs
@@ -1,4 +1,4 @@
-// Copyright 2025 PRAGMA
+// Copyright 2026 PRAGMA
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,20 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub mod has_stake_distribution;
-pub use has_stake_distribution::{HasStakeDistribution, PoolSummary};
+pub mod blocks;
+pub mod headers;
+pub mod txs;
 
-pub mod validators;
-pub use validators::*;
-
-pub mod connections;
-pub use connections::*;
-
-pub mod mempool;
-pub use mempool::*;
-
-pub mod stores;
-pub use stores::*;
-
-pub mod praos;
-pub use praos::*;
+pub use blocks::*;
+pub use headers::*;
+pub use txs::*;
